refactor: move CacheAccessor into cache.rs
