+++ src/query/storages/common/cache/src/cache.rs
@@ -16,6 +16,19 @@ use std::borrow::Borrow;
 use std::hash::Hash;
 use std::sync::Arc;
 
+pub trait CacheAccessor<K, V> {
+    fn get<Q>(&self, k: &Q) -> Option<Arc<V>>
+    where
+        K: Borrow<Q>,
+        Q: Hash + Eq + ?Sized;
+
+    fn put(&self, key: K, value: Arc<V>);
+    fn evict<Q>(&self, k: &Q) -> bool
+    where
+        K: Borrow<Q>,
+        Q: Hash + Eq + ?Sized;
+}
+
 /// The minimum interface that cache providers should implement
 pub trait StorageCache<K, V> {
     type Meter;
@@ -30,4 +43,72 @@ pub trait StorageCache<K, V> {
     where
         K: Borrow<Q>,
         Q: Hash + Eq + ?Sized;
+}
+
+mod impls {
+    use std::borrow::Borrow;
+    use std::hash::Hash;
+    use std::sync::Arc;
+
+    use parking_lot::RwLock;
+
+    use crate::cache::CacheAccessor;
+    use crate::cache::StorageCache;
+
+    impl<V, C> CacheAccessor<String, V> for Arc<RwLock<C>>
+    where C: StorageCache<String, V>
+    {
+        fn get<Q>(&self, k: &Q) -> Option<Arc<V>>
+        where
+            String: Borrow<Q>,
+            Q: Hash + Eq + ?Sized,
+        {
+            let mut guard = self.write();
+            guard.get(k).cloned()
+        }
+
+        fn put(&self, k: String, v: Arc<V>) {
+            let mut guard = self.write();
+            guard.put(k, v);
+        }
+
+        fn evict<Q>(&self, k: &Q) -> bool
+        where
+            String: Borrow<Q>,
+            Q: Hash + Eq + ?Sized,
+        {
+            let mut guard = self.write();
+            guard.evict(k)
+        }
+    }
+
+    impl<V, C> CacheAccessor<String, V> for Option<Arc<RwLock<C>>>
+    where C: StorageCache<String, V>
+    {
+        fn get<Q>(&self, k: &Q) -> Option<Arc<V>>
+        where
+            String: Borrow<Q>,
+            Q: Hash + Eq + ?Sized,
+        {
+            self.as_ref().and_then(|cache| cache.get(k))
+        }
+
+        fn put(&self, k: String, v: Arc<V>) {
+            if let Some(cache) = self {
+                cache.put(k, v);
+            }
+        }
+
+        fn evict<Q>(&self, k: &Q) -> bool
+        where
+            String: Borrow<Q>,
+            Q: Hash + Eq + ?Sized,
+        {
+            if let Some(cache) = self {
+                cache.evict(k)
+            } else {
+                false
+            }
+        }
+    }
 }
